Move Sysmon-related functions out of dynamic service helper
